Move colvarproxy_io to separate file
